requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.28.1
python-multipart==0.0.6
pydantic==2.5.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.3.5
faiss-cpu==1.7.0
numpy>=1.26.0
pandas>=2.1.1
python-jose==3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.1.1
sqlalchemy==2.0.23
pyyaml==6.0.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
langchain==0.0.350
tiktoken==0.5.2
